Update ports, doc and password source

+++ image-build.sh
@@ -7,8 +7,15 @@
 # set -x # Uncomment to debug
 # set -n # Uncomment to check script syntax without execution
 
+pass=secrets/root_password
+
+if [ ! -f ${pass} -o -z "$(cat ${pass})" ]; then 
+	echo "Bitte stelle sicher dass '${pass}' existert und den root passwort enthält" >&2
+	exit 1
+fi
+
 name=debian_server_workshop_image 
 
 docker image rm -f $name 
 
-docker build -t $name .
+docker build --secret id=root_password,src=./${pass} -t $name .
